requirements.txt
openai
python-docx
streamlit
mammoth